Make Base Fee a Float so it doesn't go 0
Fixes #20

+++ src/components/BaseFeeChart.tsx
@@ -92,7 +92,7 @@ export const BaseFeeChart = forwardRef<BaseFeeChartProps, 'div'>((props: BaseFee
       }
       
       if (chartType === 'basefee') {
-        chartData.basefee = parseInt(utils.formatUnits(block.basefee, 'gwei'))
+        chartData.basefee = parseFloat(utils.formatUnits(block.basefee, 'gwei'))
       }
 
       newData.push(chartData)
